add test case in cmd package and benchmark testing for collector package
