+++ cmd/exporter/exporter_suite_test.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestExporter(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Exporter Suite")
+}
